+++ packages/core/src/createActor.ts
@@ -70,10 +70,11 @@ const defaultOptions = {
  * its behavior based on the events it receives, which can cause effects outside
  * of the actor. When you run a state machine, it becomes an actor.
  */
-export class Actor<TLogic extends AnyActorLogic>
-  implements
-    ActorRef<SnapshotFrom<TLogic>, EventFromLogic<TLogic>, EmittedFrom<TLogic>>
-{
+export class Actor<TLogic extends AnyActorLogic> implements ActorRef<
+  SnapshotFrom<TLogic>,
+  EventFromLogic<TLogic>,
+  EmittedFrom<TLogic>
+> {
   /** The current internal state of the actor. */
   private _snapshot!: SnapshotFrom<TLogic>;
   /**
@@ -646,12 +647,14 @@ export class Actor<TLogic extends AnyActorLogic>
       }
     }
     this.observers.clear();
+    this.eventListeners.clear();
   }
   private _reportError(err: unknown): void {
     if (!this.observers.size) {
       if (!this._parent) {
         reportUnhandledError(err);
       }
+      this.eventListeners.clear();
       return;
     }
     let reportError = false;
@@ -666,6 +669,7 @@ export class Actor<TLogic extends AnyActorLogic>
       }
     }
     this.observers.clear();
+    this.eventListeners.clear();
     if (reportError) {
       reportUnhandledError(err);
     }